Rationalized view vs. edit mode names and directory organization.
