+++ src/hi/apps/collection/templates/collection/edit/panes/collection_edit_mode_panel.html
@@ -9,7 +9,7 @@
 </div>
 
 <div id="{{ DIVID.COLLECTION_EDIT_PANE }}">
-  {% include "collection/edit/panes/collection_edit.html" %}
+  {% include "collection/edit/panes/collection_properties_edit.html" %}
 </div>
 
 {% if collection_position_form %}
