+++ src/hi/apps/collection/urls.py
@@ -13,9 +13,6 @@ urlpatterns = [
              views.CollectionViewDefaultView.as_view(), 
              name='collection_view_default'),
 
-    re_path( r'^details/(?P<collection_id>\d+)$', 
-             views.CollectionDetailsView.as_view(), 
-             name='collection_details' ),
 
     re_path( r'^edit/', include('hi.apps.collection.edit.urls' )),
 ]
