+++ src/hi/apps/collection/view_mixins.py
@@ -32,7 +32,7 @@ class CollectionViewMixin:
                                   status_code           : int             = 200 ):
 
         context = collection_edit_data.to_template_context()
-        template = get_template( 'collection/edit/panes/collection_edit.html' )
+        template = get_template( 'collection/edit/panes/collection_properties_edit.html' )
         content = template.render( context, request = request )
         return antinode.response(
             insert_map = {
