+++ src/hi/apps/collection/views.py
@@ -6,11 +6,9 @@ from django.urls import reverse
 from django.views.generic import View
 
 from hi.apps.common.utils import is_ajax
-from hi.apps.location.location_manager import LocationManager
 
 from hi.enums import ViewType
 from hi.exceptions import ForceSynchronousException
-from hi.hi_async_view import HiSideView
 from hi.hi_grid_view import HiGridView
 
 from .collection_manager import CollectionManager
@@ -71,24 +69,3 @@ class CollectionViewView( HiGridView, CollectionViewMixin ):
         return context
     
     
-class CollectionDetailsView( HiSideView, CollectionViewMixin ):
-
-    def get_template_name( self ) -> str:
-        return 'collection/panes/collection_details.html'
-    
-    def should_push_url( self ):
-        return True
-    
-    def get_template_context( self, request, *args, **kwargs ):
-        collection = self.get_collection( request, *args, **kwargs )
-        
-        current_location_view = None
-        if request.view_parameters.view_type.is_location_view:
-            current_location_view = LocationManager().get_default_location_view( request = request )
-
-        collections_detail_data = CollectionManager().get_collection_details_data(
-            collection = collection,
-            location_view = current_location_view,
-            is_editing = request.view_parameters.is_editing,
-        )
-        return collections_detail_data.to_template_context()
