+++ src/hi/apps/edit/templates/edit/panes/side_header.html
@@ -1,14 +1,14 @@
 <div>
   {% if request.view_parameters.view_type.is_location_view %}
     <div class="mt-2">
-      {% include "location/edit/panes/location_edit_header.html" %}
+      {% include "location/edit/panes/location_edit_mode_header.html" %}
     </div>
     <div class="mt-2">
-      {% include "location/edit/panes/location_view_edit_header.html" %}
+      {% include "location/edit/panes/location_view_edit_mode_header.html" %}
     </div>
   {% elif request.view_parameters.view_type.is_collection %}
     <div class="mt-2">
-      {% include "collection/edit/panes/collection_edit_header.html" %}
+      {% include "collection/edit/panes/collection_edit_mode_header.html" %}
     </div>
   {% endif %}
 </div>
