+++ src/hi/apps/entity/edit/views.py
@@ -1,9 +1,10 @@
 import logging
 import re
+from typing import Any, Dict, Optional
 
 from django.core.exceptions import BadRequest, PermissionDenied
 from django.db import transaction
-from django.http import Http404
+from django.http import Http404, HttpRequest, HttpResponse
 from django.template.loader import get_template
 from django.urls import reverse
 from django.utils.decorators import method_decorator
@@ -13,16 +14,17 @@ from hi.apps.collection.collection_manager import CollectionManager
 import hi.apps.common.antinode as antinode
 from hi.apps.entity.entity_manager import EntityManager
 from hi.apps.entity.entity_pairing_manager import EntityPairingManager, EntityPairingError
+from hi.apps.entity.entity_type_transition_handler import EntityTypeTransitionHandler
 from hi.apps.entity.forms import EntityForm
 from hi.apps.entity.models import Entity, EntityPosition
 from hi.apps.entity.view_mixins import EntityViewMixin
-from hi.apps.location.location_manager import LocationManager
 from hi.apps.location.models import LocationView
+from hi.apps.location.location_manager import LocationManager
 from hi.apps.location.svg_item_factory import SvgItemFactory
 
 from hi.constants import DIVID
 from hi.decorators import edit_required
-from hi.hi_async_view import HiModalView
+from hi.hi_async_view import HiModalView, HiSideView
 
 from . import forms
 
@@ -219,3 +221,78 @@ class ManagePairingsView( HiModalView, EntityViewMixin ):
             return antinode.refresh_response()
         except EntityPairingError as epe:
             raise BadRequest( str(epe) )
+
+
+class EntityEditModeView( HiSideView, EntityViewMixin ):
+
+    def get_template_name( self ) -> str:
+        return 'entity/edit/panes/entity_edit_mode_panel.html'
+
+    def should_push_url( self ) -> bool:
+        return True
+    
+    def get_template_context( self,
+                              request : HttpRequest,
+                              *args   : Any,
+                              **kwargs: Any          ) -> Dict[str, Any]:
+        entity: Entity = self.get_entity( request, *args, **kwargs )
+
+        current_location_view = None
+        if request.view_parameters.view_type.is_location_view:
+            current_location_view = LocationManager().get_default_location_view( request = request )
+
+        entity_edit_mode_data = EntityManager().get_entity_edit_mode_data(
+            entity = entity,
+            location_view = current_location_view,
+            is_editing = request.view_parameters.is_editing,
+        )
+        return entity_edit_mode_data.to_template_context()
+
+
+class EntityPropertiesEditView( View, EntityViewMixin ):
+    """Handle entity properties editing (name, type) only - used by sidebar.
+    
+    Business logic is delegated to EntityTypeTransitionHandler following
+    the "keep views simple" design philosophy.
+    """
+
+    def post( self,
+              request : HttpRequest,
+              *args   : Any,
+              **kwargs: Any          ) -> HttpResponse:
+        entity: Entity = self.get_entity( request, *args, **kwargs )
+        
+        # Store original entity_type_str to detect changes
+        original_entity_type_str: str = entity.entity_type_str
+
+        entity_form = EntityForm( request.POST, instance = entity )
+        form_valid: bool = entity_form.is_valid()
+        
+        if form_valid:
+            # Delegate transition handling to specialized handler
+            transition_handler = EntityTypeTransitionHandler()
+            
+            transition_response: Optional[HttpResponse] = transition_handler.handle_entity_form_save(
+                request, entity, entity_form, None, original_entity_type_str
+            )
+            
+            # Now that transaction is committed, handle any transition response
+            if transition_response is not None:
+                return transition_response
+            
+            status_code: int = 200
+        else:
+            status_code: int = 400
+
+        context: Dict[str, Any] = {
+            'entity': entity,
+            'entity_form': entity_form,
+        }
+        template = get_template( 'entity/edit/panes/entity_properties_edit.html' )
+        content: str = template.render( context, request = request )
+        return antinode.response(
+            insert_map = {
+                DIVID['ENTITY_PROPERTIES_PANE']: content,
+            },
+            status = status_code,
+        )
