+++ src/hi/apps/location/edit/view_mixins.py
@@ -9,19 +9,6 @@ from hi.constants import DIVID
 
 class LocationEditViewMixin:
 
-    def location_modal_response(
-            self,
-            request                         : HttpRequest,
-            location_edit_data              : LocationEditData,
-            status_code                     : int                                = 200 ):
-        """Return modal response for full location editing (properties + attributes)"""
-        context = location_edit_data.to_template_context()
-        return antinode.modal_from_template(
-            request = request,
-            template_name = 'location/modals/location_edit.html',
-            context = context,
-        )
-
     def location_properties_response(
             self,
             request                         : HttpRequest,
@@ -38,14 +25,14 @@ class LocationEditViewMixin:
             status = status_code,
         )
         
-    def location_view_edit_response(
+    def location_view_edit_mode_response(
             self,
             request                  : HttpRequest,
             location_view_edit_data  : LocationViewEditData,
             status_code              : int                         = 200 ):
 
         context = location_view_edit_data.to_template_context()
-        template = get_template( 'location/edit/panes/location_view_edit.html' )
+        template = get_template( 'location/edit/panes/location_view_properties_edit.html' )
         content = template.render( context, request = request )
         return antinode.response(
             insert_map = {
