--- src/hi/apps/location/templates/location/edit/panes/location_edit.html
@@ -1,48 +1,0 @@
-{% load icons %}
-<form action="{% url 'location_edit_location_edit' location_id=location.id %}"
-      method="post" enctype="multipart/form-data"
-      class="form" data-async="true">
-  {% csrf_token %}
-  <div>
-    {% include "location/edit/panes/location_edit_form.html" %}
-  </div>
-  <div>
-    {% include "attribute/panes/attribute_form.html" with attribute_formset=location_attribute_formset %}
-  </div>
-  
-  <div class="text-center" >
-    <button id="hi-add-location-button" type="submit" class="btn btn-primary"
-	name="action" value="update">
-      {% icon "save" size="sm" css_class="hi-icon-left" %}UPDATE
-    </button>
-  </div>
-</form>
-
-<div class="text-center my-3" >
-  <form action="{% url 'location_attribute_upload' location_id=location.id %}"
-	method="post"
-	class="form" data-async="true">
-    {% csrf_token %}
-    <div>
-      {% include "attribute/panes/attribute_upload_form.html" with attribute_upload_form=location_attribute_upload_form  %}
-    </div>
-  </form>
-</div>
-
-{% if request.view_parameters.is_editing %}
-<div class="text-center my-3" >
-  <a href="{% url 'location_edit_svg_replace' location_id=location.id %}"
-     id="hi-svg-replace-link" class="btn btn-primary btn-control" role="button"
-     data-async="true">
-    {% icon "upload" size="sm" css_class="hi-icon-left" %}REPLACE BACKGROUND SVG IMAGE
-  </a>
-</div>
-
-<div class="my-3">
-  <a href="{% url 'location_edit_location_delete' location_id=location.id %}"
-     id="hi-delete-location-link" class="btn btn-danger btn-control" role="button"
-     data-async="true">
-    {% icon "delete" size="sm" css_class="hi-icon-left" %}DELETE LOCATION AND ALL VIEWS
-  </a>
-</div>
-{% endif %}
