+++ src/hi/apps/location/templates/location/edit/panes/location_edit_mode_panel.html
