+++ src/hi/apps/location/templates/location/edit/panes/location_view_edit_mode_form.html
