+++ src/hi/apps/location/templates/location/edit/panes/location_view_edit_mode_panel.html
@@ -8,5 +8,5 @@
   </div>
 </div>
 <div id="{{ DIVID.LOCATION_VIEW_EDIT_PANE }}">
-  {% include "location/edit/panes/location_view_edit.html" %}
+  {% include "location/edit/panes/location_view_properties_edit.html" %}
 </div>
