+++ src/hi/apps/location/templates/location/edit/panes/location_view_properties_edit.html
@@ -5,7 +5,7 @@
   {% csrf_token %}
   
   <div>
-    {% include "location/edit/panes/location_view_edit_form.html" %}
+    {% include "location/edit/panes/location_view_edit_mode_form.html" %}
   </div>
   <div class="text-center" >
     <button id="hi-add-location-view-button" type="submit" class="btn btn-primary"
